chore: Add Windows app.manifest to GTK
