+++ src/SolutionTemplate/UnoSolutionTemplate.net6/UnoSolutionTemplate.net6.csproj
@@ -107,6 +107,7 @@
     <None Include="Skia.WPF.Host\Properties\Settings.Designer.cs" />
     <None Include="Skia.Gtk\Assets\Fonts\uno-fluentui-assets.ttf" />
     <None Include="Skia.Gtk\Program.cs" />
+    <None Include="Skia.Gtk\app.manifest" />
     <None Include="Mobile\Android\Main.Android.cs" />
     <None Include="Mobile\Android\MainActivity.Android.cs" />
     <None Include="Mobile\iOS\Main.iOS.cs" />
