+++ src/SolutionTemplate/UnoSolutionTemplate/Skia.Gtk/SkiaGtk.vstemplate
@@ -21,6 +21,7 @@
         </Folder>
       </Folder>
       <ProjectItem TargetFileName="Program.cs" ReplaceParameters="true">Program.cs</ProjectItem>
+      <ProjectItem TargetFileName="app.manifest" ReplaceParameters="true">app.manifest</ProjectItem>
     </Project>
   </TemplateContent>
 </VSTemplate>
